+++ CHANGELOG.md
@@ -7,8 +7,9 @@ Features:
 * **New Resource** `google_project_usage_export_bucket` [GH-1080]
 
 IMPROVEMENTS:
-* sql: Support multiple users with the same name for different host for 1st gen SQL instances. [GH-1066]
+* compute: add support for updating alias ips in instances [GH-1084]
 * container: Fail if the ip_allocation_policy doesn't specify secondary range names [GH-1065]
+* sql: Support multiple users with the same name for different host for 1st gen SQL instances. [GH-1066]
 
 ## 1.6.0 (February 09, 2018)