+++ CHANGELOG.md
@@ -1,8 +1,9 @@
-## 1.6.1 (Unreleased)
+## 1.7.0 (Unreleased)
 
 Features:
 * **New Data Source** `google_compute_forwarding_rule` [GH-1078]
 * **New Data Source** `google_compute_vpn_gateway` [GH-1071]
+* **New Data Source** `google_storage_project_service_account` [GH-1110]
 * **New Resource** `google_folder_iam_binding` [GH-1076]
 * **New Resource** `google_folder_iam_member` [GH-1076]
 * **New Resource** `google_project_usage_export_bucket` [GH-1080]